requirements.txt
telethon==1.34.1
fastapi==0.109.1
uvicorn==0.27.0
requests==2.31.0
python-dotenv==1.0.0
